Adding script to format additional hard drives.
The script helps to format additional hard drives, so they could easily mounted after the installation .
At first the script presents a menu to select the device for the installation of coreos (all /dev/sd? devices).
In the second step you could format the remaining devices by choosing one out of a new menu.
All data and partition information on the hard drives will be deleted and a new partition ext3 formated partition will be created, that uses the whole space of the hard drive.
At last it calls the coreos-offline-install script.

./coreos-offline-install -d ${INSTALLATIONDEVICE} -i coreos-install.json -a /media/setup/app_install